+++ NEWS.md
@@ -2,6 +2,7 @@
 * Adding `pkgdown` homepage
 * Add `detritus_output` to `rcpp_nutr_output`
 * Streamlined code for `summarize_mdlrn` and `plot.mdl_rn`
+* Adding `get_production` and `plot_production` function
 
 # arrR 1.2
 * Return `pop_reserves_thres` in final `mdl_rn` object